mbi/requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
djangorestframework==3.9.0
djangorestframework-gis==0.13
idna==2.7
pkg-resources==0.0.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
